refactor(config): extract internal/config package for configuration management
Create dedicated internal/config package with no external dependencies
to avoid circular imports. This enables all packages (including keychain)
to share configuration utilities.

New internal/config package provides:
- GetConfigDir(): XDG-compliant config directory
- GetCredentialsPath(), GetTokenPath(): path helpers
- ShortenPath(): display utility
- DirName, CredentialsFile, TokenFile constants

Updated packages:
- auth: now delegates to config package
- keychain: uses config instead of duplicated configDir()

Closes #34
